Updates and fixes to cache and session.

+++ src/EpiSession_Memcached.php
@@ -1,11 +1,29 @@
 <?php
 class EpiSession_Memcached implements EpiSessionInterface
 {
+  private static $connected = false;
   private $key  = null;
   private $store= null;
 
+  public function __construct($params = array())
+  {
+    if(empty($_COOKIE[EpiSession::COOKIE]))
+    {
+      $cookieVal = md5(uniqid(rand(), true));
+      setcookie(EpiSession::COOKIE, $cookieVal, time()+1209600, '/');
+      $_COOKIE[EpiSession::COOKIE] = $cookieVal;
+    }
+    $this->host = !empty($params[0]) ? $params[0] : 'localhost';
+    $this->port = !empty($params[1]) ? $params[1] : 11211;;
+    $this->compress = !empty($params[2]) ? $params[2] : 0;;
+    $this->expiry   = !empty($params[3]) ? $params[3] : 3600;
+  }
+
   public function end()
   {
+    if(!$this->connect())
+      return;
+
     $this->memcached->delete($this->key);
     $this->store = null;
     setcookie(EpiSession::COOKIE, null, time()-86400);
@@ -13,7 +31,7 @@ class EpiSession_Memcached implements EpiSessionInterface
 
   public function get($key = null)
   {
-    if(empty($key) || !isset($this->store[$key]))
+    if(!$this->connect() || empty($key) || !isset($this->store[$key]))
       return false;
 
     return $this->store[$key];
@@ -21,12 +39,15 @@ class EpiSession_Memcached implements EpiSessionInterface
 
   public function getAll()
   {
+    if(!$this->connect())
+      return;
+
     return $this->memcached->get($this->key);
   }
 
   public function set($key = null, $value = null)
   {
-    if(empty($key))
+    if(!$this->connect() || empty($key))
       return false;
     
     $this->store[$key] = $value;
@@ -34,17 +55,29 @@ class EpiSession_Memcached implements EpiSessionInterface
     return $value;
   }
 
-  public function __construct($params = null)
+  private function connect($params = null)
   {
-    if(!empty($params))
-      $key = array_shift($params);
+    if(self::$connected)
+      return true;
 
-    if(empty($key) && empty($_COOKIE[EpiSession::COOKIE]))
-       setcookie(EpiSession::COOKIE, md5(uniqid(rand(), true)), time()+1209600, '/');
-
-    $this->memcached = EpiCache::getInstance(EpiCache::MEMCACHED);
-
-    $this->key = empty($key) ? $_COOKIE[EpiSession::COOKIE] : $key;
-    $this->store = $this->getAll();
+    if(class_exists('Memcached'))
+    {
+      $this->memcached = new Memcached;
+      if($this->memcached->addServer($this->host, $this->port))
+      {
+        self::$connected = true;
+        $this->key = empty($key) ? $_COOKIE[EpiSession::COOKIE] : $key;
+        $this->store = $this->getAll();
+        return true;
+      }
+      else
+      {
+        EpiException::raise(new EpiSessionMemcacheConnectException('Could not connect to memcache server'));
+      }
+    }
+    EpiException::raise(new EpiSessionMemcacheClientDneException('Could not connect to memcache server'));
   }
 }
+
+class EpiSessionMemcacheConnectException extends EpiException {}
+class EpiSessionMemcacheClientDneException extends EpiException {}
